Fix P8 audit findings (gas/config/devconsole(.netcode)/localization/storage/audio/gameplaytags/utils/loading/acs.persistence/ui)
Fresh-audit defects: gas periodic double-count, devconsole.netcode remote-DoS guard,
devconsole autocomplete crash, config dup-Id/dispose/coalesce, localization+gameplaytags
codegen dedup, storage atomic save, audio duck restore, utils RNG range overflow,
loading scene-cancel leak + CTS dispose, acs.persistence dup-key restore, ui HideAll callback.
All P8 [x] verified against code; regression tests added where runnable.

+++ packages/com.rubickanov.utils/Runtime/DeterministicRandom.cs
@@ -73,7 +73,7 @@ namespace Rubickanov.Utils
                 throw new ArgumentException(
                     $"maxExclusive ({maxExclusive}) must be greater than min ({min}).",
                     nameof(maxExclusive));
-            return min + (int)(Hash(a, b) % (uint)(maxExclusive - min));
+            return min + (int)(Hash(a, b) % (uint)((long)maxExclusive - min));
         }
 
         /// <summary>Returns a deterministic int in [min, maxExclusive) from three integer keys.</summary>
@@ -84,7 +84,7 @@ namespace Rubickanov.Utils
                 throw new ArgumentException(
                     $"maxExclusive ({maxExclusive}) must be greater than min ({min}).",
                     nameof(maxExclusive));
-            return min + (int)(Hash(a, b, c) % (uint)(maxExclusive - min));
+            return min + (int)(Hash(a, b, c) % (uint)((long)maxExclusive - min));
         }
 
         /// <summary>Returns a deterministic boolean (50/50) from two integer keys.</summary>
